Backend/run.py
"""
Application Entry Point
"""
from dotenv import load_dotenv
import os

# Load environment variables from .env file
load_dotenv()

from app import create_app
from config import config

# Get environment or default to development
env = os.environ.get('FLASK_ENV', 'development')
app = create_app(config.get(env, config['default']))

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5001)

